src_BRIDGE/requirements.txt
paho-mqtt>=2.0
requests>=2.31.0
